Fix MSVC warnings (#169)
Fixes #164

This fixes all remaining MSVC type conversion warnings (both x64 and x32
build) and one "code unreachable" warning (by marking
HDF5ErrMapper::ToException as [[noreturn]]).

Also, in DataSpace::getMaxDimensions it fixes the H5S_UNLIMITED ->
DataSpace::UNLIMITED replacement to work on the correct vector.

Using uint64_t for storage len and offset

+++ include/highfive/bits/H5Attribute_misc.hpp
@@ -34,7 +34,7 @@ namespace HighFive {
 inline Attribute::Attribute() {}
 
 inline size_t Attribute::getStorageSize() const {
-    return H5Aget_storage_size(_hid);
+    return static_cast<size_t>(H5Aget_storage_size(_hid));
 }
 
 inline DataType Attribute::getDataType() const {
